server/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
langchain==0.3.19
langchain_core==0.3.36
langchain_openai==0.3.6
openai==1.63.2
pydantic==2.10.6
python-dotenv==1.0.1
PyYAML==6.0.2
duckduckgo-search==7.5.1
wikipedia==1.4.0
